Replace SHA-512/128 by SHA-512/120 for exact 20-char base64
Truncating the SHA-512 to 128 bits created some problems for us
when using base64 encoding, because 128 is not exactly divisible
by the 6 bits of a base64 character.  This lead to the possibility
that different implementations would handle the padding differently
which could have the nasty consequence that lookups fail.
120 is exactly divible by both 6 and 4 and even 8, so using the
SHA-512/120 truncation avoids all such future problems of padding.